Chores/clean markers code (#1140)
* WIP

* Clean Referenced/ReferencingMarkersFromMarshalUtil

* Refactor: fixed golint errors

* Refactor: cleaned up files

* Refactor: refactored markerReferences
