+++ packages/markers/marker_test.go
@@ -1,6 +1,7 @@
 package markers
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -76,4 +77,159 @@ func TestMarkersByRank(t *testing.T) {
 	assert.Equal(t, uint64(1<<64-1), markersByRank.LowestRank())
 	assert.Equal(t, uint64(0), markersByRank.HighestRank())
 	assert.Equal(t, uint64(0), markersByRank.Size())
+}
+
+func TestReferencedMarkers(t *testing.T) {
+	referencedMarkers := NewReferencedMarkers(NewMarkers(
+		&Marker{1, 3},
+		&Marker{2, 7},
+	))
+
+	referencedMarkers.Add(8, NewMarkers(
+		&Marker{4, 9},
+	))
+
+	referencedMarkers.Add(9, NewMarkers(
+		&Marker{1, 5},
+		&Marker{2, 8},
+	))
+
+	referencedMarkers.Add(12, NewMarkers(
+		&Marker{1, 7},
+		&Marker{2, 10},
+	))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 3},
+		&Marker{2, 7},
+		&Marker{4, 9},
+	), referencedMarkers.Get(8))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 5},
+		&Marker{2, 8},
+		&Marker{4, 9},
+	), referencedMarkers.Get(10))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 5},
+		&Marker{2, 8},
+		&Marker{4, 9},
+	), referencedMarkers.Get(11))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 7},
+		&Marker{2, 10},
+		&Marker{4, 9},
+	), referencedMarkers.Get(12))
+
+	marshaledReferencedMarkers := referencedMarkers.Bytes()
+	unmarshaledReferencedMarkers, consumedBytes, err := ReferencedMarkersFromBytes(marshaledReferencedMarkers)
+	require.NoError(t, err)
+	assert.Equal(t, len(marshaledReferencedMarkers), consumedBytes)
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 3},
+		&Marker{2, 7},
+		&Marker{4, 9},
+	), unmarshaledReferencedMarkers.Get(8))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 5},
+		&Marker{2, 8},
+		&Marker{4, 9},
+	), unmarshaledReferencedMarkers.Get(10))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 5},
+		&Marker{2, 8},
+		&Marker{4, 9},
+	), unmarshaledReferencedMarkers.Get(11))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 7},
+		&Marker{2, 10},
+		&Marker{4, 9},
+	), unmarshaledReferencedMarkers.Get(12))
+
+	fmt.Println(unmarshaledReferencedMarkers)
+}
+
+func TestReferencedMarkersPanic(t *testing.T) {
+	referencedMarkers := NewReferencedMarkers(NewMarkers(
+		&Marker{1, 3},
+	))
+
+	referencedMarkers.Add(7, NewMarkers(
+		&Marker{4, 9},
+	))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 3},
+	), referencedMarkers.Get(4))
+}
+
+func TestReferencingMarkers(t *testing.T) {
+	referencingMarkers := NewReferencingMarkers()
+	referencingMarkers.Add(9, &Marker{1, 5})
+	referencingMarkers.Add(10, &Marker{3, 4})
+	referencingMarkers.Add(12, &Marker{1, 7})
+	referencingMarkers.Add(12, &Marker{2, 10})
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 5},
+		&Marker{2, 10},
+		&Marker{3, 4},
+	), referencingMarkers.Get(8))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 5},
+		&Marker{2, 10},
+		&Marker{3, 4},
+	), referencingMarkers.Get(9))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 7},
+		&Marker{2, 10},
+		&Marker{3, 4},
+	), referencingMarkers.Get(10))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 7},
+		&Marker{2, 10},
+	), referencingMarkers.Get(12))
+
+	assert.Equal(t, NewMarkers(), referencingMarkers.Get(13))
+
+	marshaledReferencingMarkers := referencingMarkers.Bytes()
+	unmarshaledReferencingMarkers, consumedBytes, err := ReferencingMarkersFromBytes(marshaledReferencingMarkers)
+	require.NoError(t, err)
+	assert.Equal(t, len(marshaledReferencingMarkers), consumedBytes)
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 5},
+		&Marker{2, 10},
+		&Marker{3, 4},
+	), unmarshaledReferencingMarkers.Get(8))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 5},
+		&Marker{2, 10},
+		&Marker{3, 4},
+	), unmarshaledReferencingMarkers.Get(9))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 7},
+		&Marker{2, 10},
+		&Marker{3, 4},
+	), unmarshaledReferencingMarkers.Get(10))
+
+	assert.Equal(t, NewMarkers(
+		&Marker{1, 7},
+		&Marker{2, 10},
+	), unmarshaledReferencingMarkers.Get(12))
+
+	assert.Equal(t, NewMarkers(), unmarshaledReferencingMarkers.Get(13))
+
+	fmt.Println(unmarshaledReferencingMarkers)
 }
