+++ packages/markers/objecstorage.go
@@ -12,9 +12,12 @@ const (
 
 	// PrefixSequenceAliasMapping defines the storage prefix for the SequenceAliasMapping object storage.
 	PrefixSequenceAliasMapping
+
+	// CacheTime defines how long objects are cached in the object storage.
+	CacheTime = 60 * time.Second
 )
 
 // objectStorageOptions contains a list of default settings for the object storage.
 var objectStorageOptions = []objectstorage.Option{
-	objectstorage.CacheTime(60 * time.Second),
+	objectstorage.CacheTime(CacheTime),
 }
